third_party/WebKit/LayoutTests/http/tests/inspector/elements/styles/selector-line-expected.txt

Tests that selector line is computed correctly regardless of its start column. Bug 110732.

Text
[expanded] 
element.style  { ()

[expanded] 
#container #inspected  { (selector-line.scss:4 -> selector-line.scss:4:5)
  color: green;

[expanded] 
#inspected  { (selector-line.html:4 -> selector-line.html:4:1)
/-- overloaded --/   color: green;

[expanded] 
div  { (user agent stylesheet)
  display: block;
